Fix depth property not working on Model tag
Fixes #924 when using `<Model enable-xr style={{'--xr-depth': '20px' }}`
the depth property was not considered when resizing the model to fit
within the container. Only the width and height property were considered
for resize to fit. If no depth is set or it's 0 then take the intrinsic
depth of the 3D model